+++ packages/sp/conf/master.d-conf/CHANGELOG.md
@@ -1,3 +1,7 @@
+201702-1
+--------
+  * Add sp-secrets package.
+
 201701-1
 --------
   * Glob paths for SPM packages.